Step 8 Complete
Added stickers that display emojis on the canvas. They are previewed similarly to the tool preview with the brush sizes, but can be placed only once before having to click on the sticker again to place another one. Currently can not undo or redo to remove the stickers, but can be cleared to get rid of it all.